fix all other kafka tests
